+++ src/Drawie.Backend.Core/Surfaces/IFramebufferInfo.cs
@@ -2,5 +2,5 @@ namespace Drawie.Backend.Core.Surfaces;
 
 public interface IFramebufferInfo
 {
-    public uint FramebufferId { get; }
+    public ulong FramebufferId { get; }
 }
